requirements.txt
fastapi
gunicorn
uvicorn
pytest
pre-commit
requests
celery
aiofiles
redis
cassandra-driver
dataparser
pandas
requests_html
selenium
python-dotenv
python-slugify
fake-useragent